defining a base inferer
