+++ clinicadl/infer/__init__.py
@@ -1,2 +1,2 @@
-from .base import Inferer
+from .abstract import Inferer
 from .simple import SimpleInferer
