+++ clinicadl/infer/simple.py
@@ -1,27 +1,16 @@
-from logging import getLogger
-from typing import Any, Optional, Sequence, TypeVar, Union, overload
+from typing import Optional, Sequence
 
 import torch
-import torch.nn as nn
-import torchio as tio
 from pydantic import Field
 
-from clinicadl.data.dataloader import Batch
-from clinicadl.data.structures import DataPoint
 from clinicadl.transforms.handlers import Postprocessing
 from clinicadl.transforms.types import TransformOrConfig
-from clinicadl.utils.config import ObjectConfig
-from clinicadl.utils.dictionary.words import OUTPUT
 from clinicadl.utils.objects import HasConfig
 
-from .base import Inferer
-
-logger = getLogger("clinicadl.early_stopping")
-T = TypeVar("T", DataPoint, Batch)
-DataPointT = TypeVar("DataPointT", bound=DataPoint)
+from .base import BaseInferer, BaseInfererConfig
 
 
-class SimpleInfererConfig(ObjectConfig["SimpleInferer"]):
+class SimpleInfererConfig(BaseInfererConfig):
     """Config class for ``SimpleInferer``."""
 
     postprocessing: Postprocessing = Field(reader=Postprocessing.from_dict)
@@ -32,7 +21,7 @@ class SimpleInfererConfig(ObjectConfig["SimpleInferer"]):
         return SimpleInferer
 
 
-class SimpleInferer(HasConfig[SimpleInfererConfig], Inferer):
+class SimpleInferer(BaseInferer, HasConfig[SimpleInfererConfig]):
     """
     For classical inference, i.e. when the whole images are passed
     in the neural network and the raw outputs are returned (with a potential
@@ -130,88 +119,7 @@ class SimpleInferer(HasConfig[SimpleInfererConfig], Inferer):
             postprocessing=postprocessing, postprocessing_on_cpu=postprocessing_on_cpu
         )
 
-    @overload
-    def __call__(
-        self,
-        x: DataPointT,
-        network: nn.Module,
-        input_dtype: Optional[torch.dtype] = None,
-        **kwargs: Any,
-    ) -> DataPointT:
-        ...
-
-    @overload
-    def __call__(
-        self,
-        x: Batch[DataPointT],
-        network: nn.Module,
-        input_dtype: Optional[torch.dtype] = None,
-        **kwargs: Any,
-    ) -> Batch[DataPointT]:
-        ...
-
-    def __call__(
-        self,
-        x: Union[DataPointT, Batch[DataPointT]],
-        network: nn.Module,
-        input_dtype: Optional[torch.dtype] = None,
-        **kwargs: Any,
-    ) -> Union[DataPointT, Batch[DataPointT]]:
-        tensor = self._get_input_tensor(x, input_dtype=input_dtype)
-
-        output = network(tensor, **kwargs)
-
-        self._add_output(x, output)
-
-        if self.config.postprocessing_on_cpu and self.config.postprocessing.transforms:
-            x.to(device="cpu")
-
-        return self._postprocess(x)
-
-    @classmethod
-    def _add_output(cls, x: Union[DataPoint, Batch], output: torch.Tensor) -> None:
-        """
-        Adds the inference output in the origin data structure.
-        """
-        if isinstance(x, DataPoint):
-            x[OUTPUT] = cls._format_output(x, output)
-        elif isinstance(x, Batch):
-            x.add_field(
-                OUTPUT, [cls._format_output(x_, out_) for x_, out_ in zip(x, output)]
-            )
-
-    @staticmethod
-    def _format_output(
-        x: DataPoint, output: torch.Tensor
-    ) -> Union[tio.Image, torch.Tensor]:
-        """
-        Formats the output, i.e. puts it in a :py:class:`torchio.Image`, or leaves it as
-        a :py:class:`torch.Tensor`.
-        """
-        try:
-            if x.label is None:
-                return tio.ScalarImage(tensor=output, affine=x.image.affine)
-            elif isinstance(x.label, tio.LabelMap):
-                return tio.LabelMap(tensor=output, affine=x.label.affine)
-        except Exception:
-            logger.info(
-                "The Inferer tried to wrap the neural network output in a torchio.Image, but an error occurred."
-            )
-
-        return output
-
-    def _postprocess(self, x: DataPointT) -> DataPointT:
-        """
-        Applies postprocessing.
-        """
-        if isinstance(x, DataPoint):
-            return self.config.postprocessing.apply(x)
-        elif isinstance(x, Batch):
-            return self.config.postprocessing.batch_apply(x)
-
-    @classmethod
-    def _from_config(cls, config):
-        return cls(
-            postprocessing=config.postprocessing.config.transforms.values,
-            **config.to_raw_dict(exclude=["postprocessing"]),
-        )  # not get_object here because we want to keep config classes as config classes
+    def _forward_pass(
+        self, tensor: torch.Tensor, network: torch.nn.Module, **kwargs
+    ) -> torch.Tensor:
+        return network(tensor, **kwargs)
